fix hang on mobile
