+++ index.html
@@ -8,6 +8,7 @@
     <link rel="stylesheet" href="./css/style.css">
     <script src="main.js" type="module" defer></script>
     <script src="./js/floorplan.js" defer></script>
+    <script src="./js/checker.js" defer></script>
     <script src="./js/modal.js" defer></script>
 </head>
 <body>
@@ -25,6 +26,7 @@
                     Floormap is recommended if you are nearby or familiar with the building. If not, it is highly encouraged to use a local map of the surrounding area.<br><br>
                     Permissions key:<br>
                     ✅ - Success<br>
+                    ⏳ - Timed out<br>
                     ❓ - Failed to retrieve<br>
                     ⛔ - Access to service denied<br>
                     ❌ - Not supported by your browser
@@ -36,11 +38,5 @@
 <div id="floorplan-container">
     <img id="floorplan" src="./assets/floorplan.jpg" alt="Floorplan">
 </div>
-<script type="module">
-    import { requestLocationPermission } from "./js/location.js";
-    import { checkNotificationPermission } from "./main.js";
-    requestLocationPermission();
-    checkNotificationPermission();
-</script>
 </body>
 </html>
