+++ js/checker.js
@@ -1,0 +1,46 @@
+export function checkAll() {
+    checkNotificationPermission();
+    checkLocationPermission();
+}
+
+function checkNotificationPermission() {
+    const notificationStatus = document.getElementById('notification');
+
+    if (Notification.permission === 'granted') {
+        notificationStatus.innerHTML = `<span class="success">Notifications ✅</span>`;
+    } else if (Notification.permission === 'denied') {
+        notificationStatus.innerHTML = `<span class="failure">Notifications ⛔</span>`;
+    } else {
+        notificationStatus.innerHTML = `<span class="failure">Notifications❓</span>`;
+    }
+}
+
+function checkLocationPermission() {
+    const locationText = document.getElementById('location');
+
+    if ('geolocation' in navigator) {
+        navigator.geolocation.getCurrentPosition(
+            () => {
+                locationText.innerHTML = `<span class="success">Location ✅</span>`;
+            }, e => {
+                if (e.code === e.PERMISSION_DENIED) {
+                    locationText.innerHTML = `<span class="failure">Location ⛔</span>`;
+                } else if (e.code === e.TIMEOUT) {
+                    locationText.innerHTML = `<span class="failure">Location ⏳</span>`;
+                } else {
+                    locationText.innerHTML = `<span class="failure">Location ❓</span>`;
+                }
+            }, {
+                timeout: 5 * 1000
+            }
+        );
+    } else {
+        locationText.innerHTML = `<span class="failure">Location ❌</span>`;
+    }
+}
+
+checkAll();
+
+document.getElementById('enabler').addEventListener('click', () => {
+    checkAll();
+});
